requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
timescaledb==0.2.0

# Optimization Engine
ortools==9.8.3296
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
prophet==1.1.4
xgboost==2.0.2
lightgbm==4.1.0
joblib==1.3.2

# Cloud Provider APIs
boto3==1.34.0
google-cloud-compute==1.14.0
azure-mgmt-compute==30.3.0
requests==2.31.0
aiohttp==3.9.1

# Infrastructure
kubernetes==28.1.0
terraform==0.0.1
docker==6.1.3

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# WebSocket & Real-time
websockets==12.0
asyncio-mqtt==0.16.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tabulate==0.9.0